ENH mostly working primitive root search

+++ cs789/lib/ExpandedEuclideanAlgorithm.h
@@ -1,6 +1,8 @@
 #ifndef TGL_EXPANDEDEUCLIDEANALGORITHM_H
 #define TGL_EXPANDEDEUCLIDEANALGORITHM_H
 
+#include "BigInteger.h"
+
 namespace tgl {
   /// Method object to use the Expanded Euclidean Algorithm
   class ExpandedEuclideanAlgorithm {
@@ -10,13 +12,13 @@ namespace tgl {
     ExpandedEuclideanAlgorithm();
   
     /// Find x0 and y0
-    void extendedGcd(int m, int n, int& x, int& y);
+    void extendedGcd(BigInteger m, BigInteger n, BigInteger& x, BigInteger& y);
 
   protected:
   private:
   
     /// Quotient & remainder
-    void divide(int m, int n, int& q, int& r);
+    void divide(BigInteger m, BigInteger n, BigInteger& q, BigInteger& r);
 
   };
 }
